% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FSW}
\alias{FSW}
\alias{FSW01}
\title{Simulated data from a respondent-driven survey of female sex workers}
\format{A data frame with the following two variables:
\describe{
\item{\code{I.object}}{a logical (TRUE/FALSE) or binary (0,1) vector
indicating whether respondents received the particular keychain prior to the
start of the respondent-driven survey}
\item{\code{Weight}}{a numeric vector of respondent-driven survey weights}
}}
\usage{
FSW
}
\description{
A total of 1000 imaginary keychains of a unique type were distributed to
imaginary female sex  workers in an imaginary urban area.  One week later a
respondent-driven survey recruited 512 imaginary respondents.  Respondents
were asked whether they received one of the imaginary keychains.  Nothing
here is real.
}
\examples{
data(FSW)
FSW[1:10, ]

}
\keyword{datasets}
