\name{RDSpopsize-package}
\alias{RDSpopsize-package}
\alias{RDSpopsize}
\docType{package}
\title{
\packageTitle{RDSpopsize}
}
\description{
\packageDescription{RDSpopsize}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RDSpopsize}
\packageIndices{RDSpopsize}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{RDSpopsize}

Maintainer: \packageMaintainer{RDSpopsize}
}
\references{
Berchenko Y, Frost SDW (2011) Capture-recapture methods and
respondent-drive sampling; their potential and limitations. Sexually
Transmitted Infections 87(4):267-268.

}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
