% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDSpopsize-package.R
\docType{package}
\name{RDSpopsize}
\alias{RDSpopsize}
\alias{RDSpopsize-package}
\title{RDSpopsize}
\description{
Estimation of the size of key populations from respondent-driven
sampling.  Currently provides only the capability to estimate population
size from two-source capture-recapture sampling or, equivalently, the
"multiplier method" commonly used in epidemiology.
}
\details{
Currently the package consists of the single function \code{RDS_2S_PSE}.
}
\author{
Steve Gutreuter: \email{sgutreuter@gmail.gov}
}
