\name{RDSpopsize-package}
\alias{RDSpopsize-package}
\alias{RDSpopsize}
\docType{package}
\title{
\packageTitle{RDSpopsize}
}
\description{
\packageDescription{RDSpopsize}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RDSpopsize}
\packageIndices{RDSpopsize}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{RDSpopsize}

Maintainer: \packageMaintainer{RDSpopsize}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
